requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
datasets==3.5.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
duplicity==0.8.21
exceptiongroup==1.2.2
fastapi==0.115.12
fasteners==0.14.1
ffmpy==0.5.0
filelock==3.16.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.12.23
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
future==0.18.2
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
gradio==5.25.2
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.8
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
joblib==1.4.2
keras==3.9.2
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
munkres==1.1.4
mysql-connector-python==9.2.0
mysqlclient==2.1.1
namex==0.0.8
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pyarrow==19.0.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.9.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.2.1
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
ruff==0.11.5
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.13.1
SecretStorage==3.3.1
semantic-version==2.10.0
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
snorkel==0.9.9
social-auth-app-django==5.4.3
social-auth-core==4.5.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.1
systemd-python==234
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.19.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
tweepy==4.14.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.1
wadllib==1.3.6
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.0
xdg==5
xkit==0.0.0
xxhash==3.5.0
yarl==1.18.3
zipp==1.0.0
